Fix grammar/mechanics/punctuation & code snippet
@pavankalyan-codes Thanks for making the previous changes! I made some extra edits for formatting purposes. Please let me know if everything looks good to you!

`Object` in the context written above should be lowercase because it is referring to an object rather than the Java `Object`.
In addition, the programming language for code snippets needs to be explicitly written.
Code snippets should be closely described! Lastly, for Docs, the first-person plural "we" is not used.