Generative AI inclusion in Contribution Guide
Hello,

I added language on our generative AI policy to the contribution guide. Please provide feedback when possible.  In particular, I struggled with whether or not to include language around using generative AI as a starting point for an entry. I'm afraid that this creates a blurred line that would be difficult to manage in the future.