Add HTML table video (#1958)
* Add HTML table video

Hi there! I've add one of Skillsoft's video walkthroughs on how to create a HTML table to the bottom of this page. I've also optimized some of the headers for SEO. I know this is the first time we've tried a YouTube embed in a Doc so I'm curious to see how it looks.

* Lint & Format

---------