[Term Entry] Python Built-in Function: .id() (#2486)
* Created id term for Python built-in function

* Apply suggestions from code review

I have added the wonderful suggestions! Thank you so much! The only thing I have to do now is to delete the second codebyte example as it was suggested that it's not needed. I will do that in VS Code and push the changes here! 😊

Co-authored-by: SSwiniarski <86081858+SSwiniarski@users.noreply.github.com>
Co-authored-by: Daksha Deep <dakshadeep1234@gmail.com>

* Update id.md 

I deleted the second codebyte example

* Apply suggestions from code review

I made additional suggested changes eliminating extra white space! 😊

Co-authored-by: Daksha Deep <dakshadeep1234@gmail.com>

* Update content/python/concepts/built-in-functions/id/id.md

Co-authored-by: Daksha Deep <dakshadeep1234@gmail.com>

* Update and rename content/python/concepts/built-in-functions/id/id.md to content/python/concepts/built-in-functions/terms/id/id.md

* review edits

---------