fix: Update codebyte example for PHP acos() method
The codebyte example prints the tag <br> along with the resultant value which can confuse the user, and serves no purpose in the example. 

A few options considered -
1) Include HTML tags and echo another example, which will serve the purpose of the <br> tag.
2) Remove the tag from the echo statement altogether.
3) Change the text from <br> tag to " Radians".

Considered option 3 as it also gave the user context of the mathematical unit of the value resulting from the method call.


Note: Since this is my first edit on Codeaademy docs, theres a good chance I might've missed some process(es). Please do correct me if that is the case.