fixed typo and added map.entrySet() explanation
Hello
This is my first commit in GitHub(part of the course). would appreciate any feedback :)
1.Fixed the typo in line 2:
From :  'HashMap is used to store items as key-value pair. The keys and values can be of either same or different types.'

To:  'HashMap is used to store items as key-value pairs. The keys and values can be of either the same or different types.'

it should be pairs, since 'items' is plural.

2. Added the map.entrySet() traversing explanation in addition to the keySet() traversing method.