[Term Entry] C++ Maps: .empty()
* Created new markdown file for the term entry on the cpp member function '.empty()'

* moved empty.md into the correct file location as specified by issue #5354

* Modified the tags.md file to include a suggestion for a tag: namely, C++. If desired, could be changed to cpp.

* Finished first complete draft of the empty.md file.

* Made edits to the empty.md file.

* Edited empty.md to delete whitespace detected by yarn:lint test

* Updated proposed change to tags.md

Updated proposed change to tags.md, for the new proposed tag to be Cpp not C++

* Update empty.md to include new proposed tag

Update empty.md to include a new proposed Cpp tag instead of a new proposed C++ tag

* Update tags.md for new proposed Cpp tag

Update tags.md for the new proposed Cpp tag to be contained in the tag list in alphabetical order

* Update empty.md

minor fixes

* Update tags.md

* Update empty.md

Deleted the 'Cpp' tag in the 'Tags' metadata section, since it was not approved to be in the tags.md file as a new tag.

* Attempted to update the lockfile by running 'yarn install'

* Revert "Attempted to update the lockfile by running 'yarn install'"

This reverts commit a9aa75e326a337764fc385c8a0ac4fa93a15fe79.

* Update empty.md

I think I found what was causing the format: verify command to fail! An extra space after the "Subjects: " header in the metadata. :)

* Minor changes

---------