Update mixedlm.md
Please, I have done the correction and also added the note as you specified. 

I will still need to exemplify that the three models can be combined in an operation, but that will be later.  I just want you to work on the aspect that is important to you and that I have already completed, then later I will do that.
 
Review the changes, and if you think I didn't still get some things right, always feel free to reach out. Thanks